+++ .eslintrc.js
@@ -35,7 +35,8 @@ module.exports = {
     'max-len': [
       'error',
       {
-        'code': 80
+        'code': 80,
+        'ignoreComments': true
       }
     ],
     'key-spacing': [