+++ src/app.js
@@ -1,5 +1,4 @@
 import express from 'express'
-import path from 'path'
 import cookieParser from 'cookie-parser'
 import logger from 'morgan'
 
@@ -14,7 +13,6 @@ app.use(logger('dev'))
 app.use(express.json())
 app.use(express.urlencoded({ extended: false }))
 app.use(cookieParser())
-app.use(express.static(path.join(__dirname, 'public')))
 
 app.use((req, res, next) => {
   res.header('Access-Control-Allow-Origin', '*')