+++ src/bin/index.js
@@ -9,34 +9,12 @@ import { connection } from '../mongo/index'
 const debug = Debug('covidperu:server')
 const PORT = process.env.PORT
 
-/**
- * Normalize a port into a number, string, or false.
- */
-
-const normalizePort = val => {
-  const port = parseInt(val, 10)
-
-  if(isNaN(port))
-    // named pipe
-    return val
-
-  if(port >= 0)
-    // port number
-    return port
-
-  return false
-}
-
-/**
- * Event listener for HTTP server "error" event.
- */
-
+// Event listener for HTTP server "error" event.
 const onError = error => {
   if(error.syscall !== 'listen')
     throw error
 
-
-  const bind = typeof port === 'string' ? 'Pipe ' + port : 'Port ' + port
+  const bind = typeof PORT === 'string' ? 'Pipe ' + PORT : 'Port ' + PORT
 
   // handle specific listen errors with friendly messages
   switch (error.code) {
@@ -53,35 +31,22 @@ const onError = error => {
   }
 }
 
-/**
- * Event listener for HTTP server "listening" event.
- */
-
+// Event listener for HTTP server "listening" event.
 const onListening = () => {
   const addr = server.address()
   const bind = typeof addr === 'string' ? 'pipe ' + addr : 'port ' + addr.port
   debug('Listening on ' + bind)
 }
 
-/**
- * Get port from environment and store in Express.
- */
+// Set the port from environment.
+app.set('port', PORT)
+console.log(`APP RUNNING AT PORT ${PORT}`)
 
-const port = normalizePort(PORT || '3000')
-app.set('port', port)
-console.log(`APP RUNNING AT PORT ${port}`)
-
-/**
- * Create HTTP server.
- */
-
+// Create HTTP server.
 const server = http.createServer(app)
 
-/**
- * Listen on provided port, on all network interfaces.
- */
-
-server.listen(port)
+// Listen on provided port, on all network interfaces.
+server.listen(PORT)
 server.on('error', onError)
 server.on('listening', onListening)
 
@@ -94,6 +59,7 @@ connection.once('open', () => {
   console.log('We are connected with the database!')
 })
 
+// Setting the cronjob for production
 const job = new CronJob('00 00 21 * * *', () => {
   const currentDate = new Date(new Date().getTime() + 24*60*60*1000)
   let month, day