+++ src/functions/imgToText.js
@@ -1,0 +1,22 @@
+import ocrSpaceApi from 'ocr-space-api-alt2'
+
+const OCR_API_KEY = process.env.OCR_API_KEY
+
+const options = {
+  apikey  : `<${OCR_API_KEY}>`,
+  filetype: 'png'
+}
+
+const getText = async path => {
+  try {
+    let result = await ocrSpaceApi(path, options)
+
+    return result
+  } catch (error) {
+    throw new Error(
+      `There was an error trying to get text from the image.\n${error.message}`
+    )
+  }
+}
+
+export { getText }